core: make the jailhouse-header arch-independent
The header of jailhouse is defined with arch-dependent types such as
`unsigned long` which on linux varies in size. Because this header can
be considered the "communication"-relay with the environment, it should
be type-safe on any arch. Thus change all types into fixed-size-types.

[Jan: adjusted jailhouse_entry to keep unsigned int]

+++ hypervisor/setup.c
@@ -206,5 +206,5 @@ hypervisor_header = {
 	.signature = JAILHOUSE_SIGNATURE,
 	.core_size = (unsigned long)__hv_core_end - JAILHOUSE_BASE,
 	.percpu_size = sizeof(struct per_cpu),
-	.entry = arch_entry,
+	.entry = (unsigned long)arch_entry,
 };
